+++ services/encyclopedias/equipments.go
@@ -7,7 +7,7 @@ import (
 	"github.com/kaellybot/kaelly-encyclopedia/models/mappers"
 )
 
-func (service *Impl) getEquipmentByID(ctx context.Context, id int32, correlationID,
+func (service *Impl) getEquipmentByID(ctx context.Context, id int64, correlationID,
 	lg string) (*amqp.EncyclopediaItemAnswer, error) {
 	equipment, err := service.sourceService.GetEquipmentByID(ctx, id, lg)
 	if err != nil {
