+++ services/encyclopedias/sets.go
@@ -11,7 +11,7 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-func (service *Impl) getSetByID(ctx context.Context, id int32, correlationID,
+func (service *Impl) getSetByID(ctx context.Context, id int64, correlationID,
 	lg string) (*amqp.EncyclopediaItemAnswer, error) {
 	set, err := service.sourceService.GetSetByID(ctx, id, lg)
 	if err != nil {
@@ -19,7 +19,7 @@ func (service *Impl) getSetByID(ctx context.Context, id int32, correlationID,
 	}
 
 	items := service.getSetEquipments(ctx, set, correlationID, lg)
-	icon := service.getSetIcon(set.GetAnkamaId())
+	icon := service.getSetIcon(int64(set.GetAnkamaId()))
 	return mappers.MapSet(set, items, icon, service.equipmentService), nil
 }
 
@@ -31,13 +31,13 @@ func (service *Impl) getSetByQuery(ctx context.Context, query, correlationID,
 	}
 
 	items := service.getSetEquipments(ctx, set, correlationID, lg)
-	icon := service.getSetIcon(set.GetAnkamaId())
+	icon := service.getSetIcon(int64(set.GetAnkamaId()))
 	return mappers.MapSet(set, items, icon, service.equipmentService), nil
 }
 
 func (service *Impl) getSetEquipments(ctx context.Context, set *dodugo.EquipmentSet, correlationID,
 	lg string) map[int32]*dodugo.Weapon {
-	var getItemByID func(ctx context.Context, equipmentID int32, lg string) (*dodugo.Weapon, error)
+	var getItemByID func(ctx context.Context, equipmentID int64, lg string) (*dodugo.Weapon, error)
 	if set.GetIsCosmetic() {
 		getItemByID = service.sourceService.GetCosmeticByID
 	} else {
@@ -46,7 +46,7 @@ func (service *Impl) getSetEquipments(ctx context.Context, set *dodugo.Equipment
 
 	items := make(map[int32]*dodugo.Weapon)
 	for _, itemID := range set.GetEquipmentIds() {
-		item, errItem := getItemByID(ctx, itemID, lg)
+		item, errItem := getItemByID(ctx, int64(itemID), lg)
 		if errItem != nil {
 			log.Error().Err(errItem).
 				Str(constants.LogCorrelationID, correlationID).
@@ -60,7 +60,7 @@ func (service *Impl) getSetEquipments(ctx context.Context, set *dodugo.Equipment
 	return items
 }
 
-func (service *Impl) getSetIcon(setID int32) string {
+func (service *Impl) getSetIcon(setID int64) string {
 	setDB, found := service.setService.GetSetByDofusDude(setID)
 	if found {
 		return setDB.Icon
